# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: BlockMesh API
  description: Blockchain RPC gateway with multi-tenant support, rate limiting, caching, and admin management.
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Gateway
  - url: http://localhost:8081
    description: Admin

tags:
  - name: Gateway
    description: Public RPC proxy and health endpoints
  - name: Admin — Health
    description: Admin service health checks
  - name: Admin — Stats
    description: Usage analytics and dashboards
  - name: Admin — Tenants
    description: Tenant (API key) management
  - name: Admin — Blockchain
    description: Blockchain network configuration
  - name: Admin — Blocks
    description: Block ingestion data

paths:
  # Gateway
  /v1/:
    post:
      tags: [Gateway]
      summary: Proxy JSON-RPC request
      description: |
        Proxies any Ethereum JSON-RPC method to the configured blockchain network.
        Authentication via `Authorization: Bearer <api_key>` header.
        Supports caching for common methods (eth_chainId, eth_blockNumber, etc.).
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                jsonrpc:
                  type: string
                  example: "2.0"
                method:
                  type: string
                  example: eth_getBalance
                params:
                  type: array
                  items: {}
                  example: ["0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb", "latest"]
                id:
                  type: integer
                  example: 1
      responses:
        "200":
          description: JSON-RPC response (may contain RPC-level error)
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Missing or invalid API key
        "429":
          description: Rate limit exceeded
        "502":
          description: Upstream blockchain node unavailable
        "503":
          description: No blockchain network configured

  /health/nodes:
    get:
      tags: [Gateway]
      summary: Node health status
      description: Returns health and latency for all configured blockchain endpoints. RPC URLs are redacted.
      responses:
        "200":
          description: List of network health statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NetworkHealth"

  # Admin — Health
  /health:
    get:
      tags: [Admin — Health]
      summary: Admin service health
      security:
        - adminSecret: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "403":
          description: Invalid or missing admin secret

  # Admin — Stats 
  /stats/summary:
    get:
      tags: [Admin — Stats]
      summary: Request statistics summary
      security:
        - adminSecret: []
      parameters:
        - name: range
          in: query
          schema:
            type: string
            enum: [15m, 1h, 24h]
            default: 1h
      responses:
        "200":
          description: Aggregated statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"
        "400":
          description: Invalid range parameter
        "403":
          description: Forbidden

  # Admin — Tenants
  /tenants:
    get:
      tags: [Admin — Tenants]
      summary: List all tenants
      security:
        - adminSecret: []
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"

    post:
      tags: [Admin — Tenants]
      summary: Create a new tenant
      security:
        - adminSecret: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantRequest"
      responses:
        "201":
          description: Tenant created with API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantCreated"
        "400":
          description: Invalid request or no blockchain network available
        "403":
          description: Forbidden

  /tenants/{tenantId}:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Admin — Tenants]
      summary: Get tenant by ID
      security:
        - adminSecret: []
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Tenant not found
        "403":
          description: Forbidden

    put:
      tags: [Admin — Tenants]
      summary: Update tenant
      security:
        - adminSecret: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantRequest"
      responses:
        "200":
          description: Update confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
        "404":
          description: Tenant not found
        "403":
          description: Forbidden

    delete:
      tags: [Admin — Tenants]
      summary: Delete tenant
      security:
        - adminSecret: []
      responses:
        "200":
          description: Deletion confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        "404":
          description: Tenant not found
        "403":
          description: Forbidden

  /tenants/{tenantId}/rotate-key:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Admin — Tenants]
      summary: Rotate tenant API key
      security:
        - adminSecret: []
      responses:
        "200":
          description: New API key generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key:
                    type: string
        "404":
          description: Tenant not found
        "403":
          description: Forbidden

  /tenants/{tenantId}/usage:
    parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: day
        in: query
        schema:
          type: string
          format: date
        description: Date in YYYY-MM-DD format. Defaults to today.
    get:
      tags: [Admin — Tenants]
      summary: Get daily usage for a tenant
      security:
        - adminSecret: []
      responses:
        "200":
          description: Usage records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usage"
        "400":
          description: Invalid date format
        "404":
          description: Tenant not found
        "403":
          description: Forbidden

  # Admin — Blockchain 
  /blockchain:
    get:
      tags: [Admin — Blockchain]
      summary: List blockchain networks
      security:
        - adminSecret: []
      responses:
        "200":
          description: List of blockchain configs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockchainConfig"

    post:
      tags: [Admin — Blockchain]
      summary: Add a blockchain network
      security:
        - adminSecret: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlockchainConfigRequest"
      responses:
        "201":
          description: Network created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockchainConfig"
        "400":
          description: Missing required fields
        "403":
          description: Forbidden

  /blockchain/{networkId}:
    parameters:
      - name: networkId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Admin — Blockchain]
      summary: Get blockchain network by ID
      security:
        - adminSecret: []
      responses:
        "200":
          description: Network config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockchainConfig"
        "404":
          description: Network not found
        "403":
          description: Forbidden

    put:
      tags: [Admin — Blockchain]
      summary: Update blockchain network
      security:
        - adminSecret: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlockchainConfigRequest"
      responses:
        "200":
          description: Update confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
        "404":
          description: Network not found
        "403":
          description: Forbidden

    delete:
      tags: [Admin — Blockchain]
      summary: Delete blockchain network
      security:
        - adminSecret: []
      responses:
        "200":
          description: Deletion confirmed
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        "403":
          description: Forbidden

  /blockchain/test:
    post:
      tags: [Admin — Blockchain]
      summary: Test RPC endpoint connectivity
      description: Attempts eth_chainId call against provided endpoint(s).
      security:
        - adminSecret: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rpc_endpoint_1:
                  type: string
                  format: uri
                rpc_endpoint_2:
                  type: string
                  format: uri
              required: [rpc_endpoint_1]
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                  chain_id:
                    type: string
                  error:
                    type: string
        "400":
          description: Invalid request
        "403":
          description: Forbidden

  # Admin — Blocks
  /blocks:
    get:
      tags: [Admin — Blocks]
      summary: List recent ingested blocks
      description: Returns the 50 most recent blocks across all networks.
      security:
        - adminSecret: []
      responses:
        "200":
          description: List of blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Block"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Tenant API key (bm_live_...)
    adminSecret:
      type: apiKey
      in: header
      name: X-Admin-Secret
      description: Admin authentication secret

  schemas:
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        quota_rpm:
          type: integer
        quota_rps:
          type: integer
        quota_daily:
          type: integer
        plan:
          type: string
          example: free
        blockchain_network_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    CreateTenantRequest:
      type: object
      properties:
        name:
          type: string
        quota_rpm:
          type: integer
        quota_rps:
          type: integer
        quota_daily:
          type: integer
        plan:
          type: string
        blockchain_network_id:
          type: string
          format: uuid

    UpdateTenantRequest:
      type: object
      properties:
        name:
          type: string
        quota_rpm:
          type: integer
        quota_rps:
          type: integer
        quota_daily:
          type: integer
        plan:
          type: string
        blockchain_network_id:
          type: string
          format: uuid

    TenantCreated:
      allOf:
        - $ref: "#/components/schemas/Tenant"
        - type: object
          properties:
            api_key:
              type: string

    Usage:
      type: object
      properties:
        tenant_id:
          type: string
          format: uuid
        method:
          type: string
        count:
          type: integer
        bytes_in:
          type: integer
        period:
          type: string
          format: date-time

    BlockchainConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        rpc_endpoint_1:
          type: string
          format: uri
        rpc_endpoint_2:
          type: string
          format: uri
        chain_id:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CreateBlockchainConfigRequest:
      type: object
      properties:
        name:
          type: string
        rpc_endpoint_1:
          type: string
          format: uri
        rpc_endpoint_2:
          type: string
          format: uri
        chain_id:
          type: string
        enabled:
          type: boolean

    Block:
      type: object
      properties:
        number:
          type: integer
        hash:
          type: string
        parent_hash:
          type: string
        timestamp:
          type: string
          format: date-time
        tx_count:
          type: integer
        network_id:
          type: string
          format: uuid
        network_name:
          type: string

    NetworkHealth:
      type: object
      properties:
        network_id:
          type: string
        nodes:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                description: Redacted endpoint host
              healthy:
                type: boolean
              latency_ms:
                type: integer
              last_check:
                type: string
                format: date-time
              consecutive_fails:
                type: integer
              total_requests:
                type: integer
              total_failures:
                type: integer

    StatsSummary:
      type: object
      properties:
        range:
          type: string
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        totals:
          type: object
          properties:
            requests:
              type: integer
            success:
              type: integer
            errors:
              type: integer
            cache_hits:
              type: integer
            cache_misses:
              type: integer
        latency:
          type: object
          properties:
            avg_ms:
              type: number
            p95_ms:
              type: number
        top_methods:
          type: array
          items:
            $ref: "#/components/schemas/StatsCount"
        top_statuses:
          type: array
          items:
            $ref: "#/components/schemas/StatsCount"
        top_networks:
          type: array
          items:
            $ref: "#/components/schemas/StatsCount"
        series:
          type: array
          items:
            type: object
            properties:
              time:
                type: string
                format: date-time
              requests:
                type: integer
              errors:
                type: integer
              cache_hits:
                type: integer

    StatsCount:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer